+++ src/domains/auth/naaClient.ts
@@ -6,7 +6,9 @@ let pcaPromise: Promise<IPublicClientApplication> | null = null
 export const naaInit = () => {
   if (!pcaPromise) {
     pcaPromise = (async () => {
-      try { await teamsApp.initialize() } catch {}
+      try { await teamsApp.initialize() } catch {
+        // ignore
+      }
       const clientId = import.meta.env.VITE_CLIENT_ID as string
       const tenantId = (import.meta.env.VITE_TENANT_ID as string) || 'fd656490-ea47-45d1-a9a2-d102f4d92017'
       const authority = `https://login.microsoftonline.com/${tenantId}`
@@ -30,6 +32,7 @@ const tokenCache: Record<string, { token: string; exp: number }> = {}
 const isExpired = (exp: number, skew = 60) => Math.floor(Date.now() / 1000) + skew >= exp
 
 export const acquireNaaToken = async (scopes: string[]): Promise<string> => {
+  console.log('acquireNaaToken', scopes)
   const key = scopes.sort().join(' ')
   const cached = tokenCache[key]
   if (cached && !isExpired(cached.exp)) return cached.token
